backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.2
httpx==0.25.2
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
ipaddress==1.0.23
dnspython==2.4.2
validators==0.22.0
aiofiles==23.2.1